feat: add timebar examples
