+++ examples/tool/timebar/index.zh.md
@@ -7,4 +7,4 @@ G6 中内置的 TimeBar 组件。
 
 ## 使用指南
 
-下面的代码演示展示了如何在图上使用 TimeBar。TimeBar 的样式可以参考 demo 2 进行配置。
+下面的代码演示展示了如何在图上使用 TimeBar。目前仅提供了三种类型的默认时间轴组件，更多配置的时间轴组件敬请期待。
