+++ gatsby-browser.js
@@ -1,4 +1,4 @@
-// window.g6 = require('./src/index.ts'); // import the source for debugging
-window.g6 = require('./dist/g6.min.js'); // import the package for webworker
+window.g6 = require('./src/index.ts'); // import the source for debugging
+// window.g6 = require('./dist/g6.min.js'); // import the package for webworker
 window.insertCss = require('insert-css');
 window.Chart = require('@antv/chart-node-g6');
